[Fix] Migration files data type fixed

+++ db/migrate/20211212113128_create_chats.rb
@@ -1,8 +1,8 @@
 class CreateChats < ActiveRecord::Migration[5.2]
   def change
     create_table :chats do |t|
-      t.integer :user_id
-      t.integer :room_id
+      t.bigint :user_id
+      t.bigint :room_id
       t.text :message
 
       t.timestamps
