+++ db/migrate/20211212113305_create_user_rooms.rb
@@ -1,8 +1,8 @@
 class CreateUserRooms < ActiveRecord::Migration[5.2]
   def change
     create_table :user_rooms do |t|
-      t.integer :user_id
-      t.integer :room_id
+      t.bigint :user_id
+      t.bigint :room_id
 
       t.timestamps
     end
